+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project does **not** adhere to [Semantic Versioning](https://semver.org
 
 ## [Unreleased]
 
+### Fixed
+
+- On longer pages, messages were shown outside the viewport.
+
 ## [1.0.0-rc.2] - 2026-08-13
 
 ### Added